pyproject.toml
[project]
name = "programming-habits"
version = "1.0.0"
description = "Productivity timer with health breaks for Linux"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dbus-python>=1.4.0",
    "pyinstaller>=6.16.0",
    "pyqt6>=6.10.0",
]
